+++ routes/index.js
@@ -1,0 +1,9 @@
+const express = require('express');
+
+const api = require('./api');
+
+const app = express();
+
+app.use('/api', api);
+
+module.exports = app;
